CW Issue #768: Switch bind to use cwctl
