+++ dev/org.eclipse.codewind.core/META-INF/MANIFEST.MF
@@ -19,6 +19,7 @@ Export-Package: io.socket.client,
  io.socket.emitter,
  org.eclipse.codewind.core,
  org.eclipse.codewind.core.internal,
+ org.eclipse.codewind.core.internal.cli,
  org.eclipse.codewind.core.internal.connection,
  org.eclipse.codewind.core.internal.console,
  org.eclipse.codewind.core.internal.constants,
