+++ dev/org.eclipse.codewind.core/src/org/eclipse/codewind/core/internal/cli/InstallStatus.java
@@ -9,8 +9,9 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package org.eclipse.codewind.core.internal;
+package org.eclipse.codewind.core.internal.cli;
 
+import org.eclipse.codewind.core.internal.Logger;
 import org.eclipse.codewind.core.internal.connection.CodewindConnection;
 import org.json.JSONArray;
 import org.json.JSONException;
