+++ dev/org.eclipse.codewind.core/src/org/eclipse/codewind/core/internal/messages/messages.properties
@@ -73,6 +73,7 @@ StopCodewindJobLabel=Stopping Codewind
 RemovingCodewindJobLabel=Removing Codewind docker images
 CodewindStatusJobLabel=Retrieving Codewind status
 CreateProjectTaskLabel=Creating project: {0}
+BindingProjectTaskLabel=Binding project: {0}
 ValidateProjectTaskLabel=Validating project: {0}
 DeleteProjectJobLabel=Deleting project contents: {0}
 DeleteProjectError=There may be running operations that are holding on to resources. Wait for operations to complete and then try deleting {0} from the Project Explorer view.
