+++ dev/org.eclipse.codewind.test/META-INF/MANIFEST.MF
@@ -13,6 +13,7 @@ Import-Package: io.socket.client,
  io.socket.emitter,
  junit.framework,
  org.eclipse.codewind.core.internal,
+ org.eclipse.codewind.core.internal.cli,
  org.eclipse.codewind.core.internal.connection,
  org.eclipse.codewind.core.internal.console,
  org.eclipse.codewind.core.internal.constants,
